Move tests into ExclusiveGateway spec
